patch 0.4.2: FreeRTOS-based firmware.
This is the first version of the Xling firmware based on the real-time
kernel of FreeRTOS v10.2.1.

There is only one "display" task up and running besides the idle one,
provided by the kernel.

There is an ADC configured to measure a Li-Ion battery voltage and a
status pin to control status of the battery.

The SH1106 display driver has been optimized to minimize a time spent in
the SPI interrupt.

Signed-off-by: Dmitry Salychev <dsl@mcusim.org>