patch 0.4.6: Draft of the sleep mode task.
Problem: Xling consumes a lot of power without putting its display and
MCU into sleep mode. The full battery is enough to power it during
24-26 hours only.

Solution: A separate "sleep mode" task which is responsible to monitor
the device activity (no buttons pressed or battery charging started)
within a selected period of time. If there is no activity, all of the
tasks including the sleep mode one will be suspended in order to let the
FreeRTOS scheduler know that the low-power tickless mode can be
activated.

Signed-off-by: Dmitry Salychev <dsl@mcusim.org>