patch 0.4.8: An initial support of low power mode.
An initial implementation of the portSUPPRESS_TICKS_AND_SLEEP()
macro has been written for the FreeRTOS port and ATmega1284P MCU.

The microcontroller will be switched to the sleep mode when the Idle
task is the onle task able to run and at least further
configEXPECTED_IDLE_TIME_BEFORE_SLEEP ticks have passed.

The next step is to implement the Sleep Mode task to suspend all of the
tasks gracefully.

Signed-off-by: Dmitry Salychev <dsl@mcusim.org>